+++ browser/components/UserTable.tsx
@@ -56,6 +56,11 @@ export default class UserTable extends React.Component<Props, State> {
               accessor: 'company_alias',
             },
             {
+              Header: <b>GitHub Alias</b>,
+              accessor: 'github_alias',
+              Cell: cellInfo => <a href={'https://github.com/' + cellInfo.value} target="_blank"> {cellInfo.value} </a>
+            },
+            {
               Header: <b># Contribs Last Week</b>,
               accessor: 'contribWeek',
             },