+++ server/api/strategicgroups/index.ts
@@ -161,6 +161,7 @@ export async function getStrategicProject(req, id) {
   }
 
   for (const user of users) {
+    const github_alias = (await dbusers.getGitHubAliasByUsername(user)).github_alias;
     const contribWeek = await dbcontributions.getLastWeekCount(projId, [user]);
     const contribMTD = await dbcontributions.getMTDCount(projId, [user]);
     const contribMonth = await dbcontributions.getLastMonthCount(projId, [
@@ -174,6 +175,7 @@ export async function getStrategicProject(req, id) {
       contribMTD: parseInt(contribMTD.numcontribs, 10),
       contribMonth: parseInt(contribMonth.numcontribs, 10),
       contribYear: parseInt(contribYear.numcontribs, 10),
+      github_alias
     };
     userList.push(data);
   }